refactor: hoist duplicated PyRuleEntry into python_receipt_common
`PyRuleEntry` (the Python-side item-category rule mirror: keywords/category/
tags/priority) plus its `FromPyObject` impl were copy-pasted byte-identically
across three wrapper modules — python_receipt_parser, python_receipt_categories,
python_receipt_staged_json — and each then repeated the same
`PyRuleEntry -> receipt_categories::CategoryRule` field-by-field map.

Define it once in python_receipt_common and add `From<PyRuleEntry> for
CategoryRule` so the three conversions collapse to `.map(CategoryRule::from)`.
Net -36 lines; no behavior change.

Deliberately left untouched (verified NOT true duplicates):
- PyRuleLayersInput — parser uses Vec<(String,String)>/ParserRuleLayers,
  categories uses HashMap/CategoryRuleLayers with extra validation.
- PyBboxInput — same shape but parser's FromPyObject is lenient (missing→0.0)
  while spatial's is strict (missing→error); merging would change behavior.

Verified: cargo check --all-targets; test_item_categories.py +
test_receipt_core_parity.py (66 passed).
