+++ src/python_receipt_parser.rs
@@ -10,26 +10,7 @@ use receipt_core::receipt_parse_helpers;
 use receipt_core::receipt_parser;
 use receipt_core::receipt_spatial;
 
-#[derive(Clone, Debug)]
-struct PyRuleEntry {
-    keywords: Vec<String>,
-    category: Option<String>,
-    tags: Vec<String>,
-    priority: i32,
-}
-
-impl<'a, 'py> FromPyObject<'a, 'py> for PyRuleEntry {
-    type Error = PyErr;
-
-    fn extract(ob: Borrowed<'a, 'py, PyAny>) -> Result<Self, Self::Error> {
-        Ok(Self {
-            keywords: ob.getattr("keywords")?.extract::<Vec<String>>()?,
-            category: ob.getattr("category")?.extract::<Option<String>>()?,
-            tags: ob.getattr("tags")?.extract::<Vec<String>>()?,
-            priority: ob.getattr("priority")?.extract::<i32>()?,
-        })
-    }
-}
+use crate::python_receipt_common::PyRuleEntry;
 
 #[derive(Clone, Debug)]
 struct PyRuleLayersInput {
@@ -73,12 +54,7 @@ fn to_rule_layers(input: PyRuleLayersInput) -> receipt_parser::ParserRuleLayers 
             rules: input
                 .rules
                 .into_iter()
-                .map(|rule| receipt_categories::CategoryRule {
-                    keywords: rule.keywords,
-                    category: rule.category,
-                    tags: rule.tags,
-                    priority: rule.priority,
-                })
+                .map(receipt_categories::CategoryRule::from)
                 .collect(),
             exact_only_keywords: input.exact_only_keywords,
             account_mapping: category_account_mapping,
